musicplayer code refactor
1. code refactor, decouples the components;
